pyproject.toml
# SPDX-FileCopyrightText: Copyright © 2026 Idiap Research Institute <contact@idiap.ch>
# SPDX-FileContributor: William Droz <william.droz@idiap.ch>
# SPDX-License-Identifier: GPL-3.0-only

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[project]
name = "aidays2026workshop"
version = "1.0.0"
description = "AI Days 2026 Workshop - First step with Agents"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "William Droz", email = "william.droz@idiap.ch" },
]

dependencies = [
    "browser-use==0.11.13",
    "fastmcp~=3.1.0",
    "grimoireplot==0.0.2",
    "loguru==0.7.3",
    "nicegui~=3.8.0",
    "pandas~=3.0.1",
    "polars~=1.38.1",
    "pydantic-ai~=1.61.0",
    "python-dotenv~=1.2.2",
    "smolagents[toolkit]~=1.24.0",
]

[project.optional-dependencies]
scraping = [
    "playwright~=1.58.0",
]

[dependency-groups]
dev = [
    "pytest~=9.0.2",
]
docs = [
    "mkslides~=2.0.15",
]


[tool.pytest.ini_options]
testpaths = ["tests"]

[project.scripts]
plot_mcp = "aidays2026workshop.plot_mcp:main"
